fix: preserve the model activation dtype
Cast the FP32 trainable steering vector at the residual hook boundary, matching inference-time AdditiveSteering behavior and preventing BF16 model layers from receiving promoted FP32 activations.

+++ src/neural_steering/melbo.py
@@ -149,7 +149,7 @@ class MELBO:
             target_activations = activations
 
         def add_steering_vector(activations, hook):
-            return activations + steering_vector
+            return activations + steering_vector.to(activations.dtype)
 
         with torch.no_grad(), self.model.hooks(
             fwd_hooks=[(target_hook, capture_target)]
